Moved the header bar for the editor into the doceditor file
